improved effeciency
skip download if already exist, skip extract if already extracted

+++ download_debloater.sh
@@ -4,9 +4,10 @@ set -e
 
 source config.sh
 
+
+
 bash check_internet.sh
 
-###if [ ! -z $curl ]; then
 echo "Getting debloater latest release version..."
 debloater_latest_release_version=$(curl -sL $debloater_release_page_url | grep -oE '[0-9]+\.[0-9]+' | sed -n '1p')
 echo "latest release version: $debloater_latest_release_version"
@@ -22,8 +23,20 @@ fi
 
 cd $working_dir
 
-echo "Downloading $debloater_download_url"
-curl -O -L $debloater_download_url
+# download debloater if latest not already exist
+if [ ! -f "$debloater_tar_file_name" ]; then
+    echo "Downloading $debloater_download_url"
+    curl -O -L $debloater_download_url
+else
+    echo "latest debloater already exist. Skipping download."
+fi
 
-echo "Extracting debloater tar file..."
-tar -xvf $debloater_tar_file_name
+folder_name=$(find . -maxdepth 1 -type d -name "*${debloater_latest_release_version}*")
+
+# extract debloater if latest not already extracted
+if [ ! -d "$folder_name" ]; then
+    echo "Extracting debloater tar file..."
+    tar -xvf $debloater_tar_file_name
+else
+    echo "latest debloater already extracted. Skipping extract."
+fi
